ref: IP取得処理の共通化
