+++ server/utils/feedAnalytics.ts
@@ -1,4 +1,5 @@
 import { prisma } from "./prisma";
+import { getClientIP } from "./ip";
 
 /**
  * RSSリーダーの種類を判別する
@@ -60,32 +61,6 @@ export function detectRSSReader(userAgent: string | null | undefined): string {
   return "Other";
 }
 
-/**
- * IPアドレスを取得する（プロキシ考慮）
- */
-export function getClientIP(event: any): string {
-  const headers = getHeaders(event);
-
-  // プロキシヘッダーをチェック
-  const forwardedFor = headers["x-forwarded-for"];
-  if (forwardedFor) {
-    const ips = forwardedFor.split(",");
-    return ips[0]?.trim() || "unknown";
-  }
-
-  const realIP = headers["x-real-ip"];
-  if (realIP) {
-    return realIP;
-  }
-
-  const cfConnectingIP = headers["cf-connecting-ip"];
-  if (cfConnectingIP) {
-    return cfConnectingIP;
-  }
-
-  // フォールバック
-  return event.node?.req?.socket?.remoteAddress || "unknown";
-}
 
 /**
  * フィードアクセスログを非同期で記録
